[lab5] - connected authorization db, created rules

+++ src/main/java/pl/edu/pg/eti/kask/blog/comment/service/CommentService.java
@@ -4,11 +4,18 @@ import lombok.NoArgsConstructor;
 import pl.edu.pg.eti.kask.blog.article.repository.ArticleRepository;
 import pl.edu.pg.eti.kask.blog.comment.entity.Comment;
 import pl.edu.pg.eti.kask.blog.comment.repository.CommentRepository;
+import pl.edu.pg.eti.kask.blog.user.entity.User;
+import pl.edu.pg.eti.kask.blog.user.entity.UserRoles;
+import pl.edu.pg.eti.kask.blog.user.service.UserService;
 
+import javax.annotation.security.RolesAllowed;
+import javax.ejb.EJBAccessException;
 import javax.ejb.LocalBean;
 import javax.ejb.Stateless;
 import javax.inject.Inject;
+import javax.security.enterprise.SecurityContext;
 import java.io.Serializable;
+import java.security.Principal;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,15 +26,26 @@ import java.util.Optional;
 @Stateless
 @LocalBean
 @NoArgsConstructor
+@RolesAllowed(UserRoles.USER)
 public class CommentService implements Serializable {
 
     private CommentRepository commentRepository;
     private ArticleRepository articleRepository;
+    private SecurityContext securityContext;
+
+
+    private UserService userService;
 
     @Inject
-    public CommentService(CommentRepository commentRepository, ArticleRepository articleRepository) {
+    public void setUserService(UserService userService) {
+        this.userService = userService;
+    }
+
+    @Inject
+    public CommentService(CommentRepository commentRepository, ArticleRepository articleRepository, SecurityContext securityContext) {
         this.commentRepository = commentRepository;
         this.articleRepository = articleRepository;
+        this.securityContext = securityContext;
     }
 
     /**
@@ -46,9 +64,12 @@ public class CommentService implements Serializable {
      * @param comment data of comment to be created
      */
     public void createComment(Comment comment) {
+        User user = userService.findByName(securityContext.getCallerPrincipal().getName());
+        comment.setUser(user);
+        user.getComments().add(comment);
         commentRepository.create(comment);
-        articleRepository.findById(comment.getArticle().getId())
-                .ifPresent(article -> article.getComments().add(comment));
+//        articleRepository.findById(comment.getArticle().getId())
+//                .ifPresent(article -> article.getComments().add(comment));
     }
 
     /**
@@ -56,8 +77,15 @@ public class CommentService implements Serializable {
      *
      * @param comment comment to be deleted
      */
-    public void delete(Comment comment) {
-        commentRepository.delete(comment);
+    public void delete(Comment comment) throws EJBAccessException {
+        Comment originalComment = commentRepository.findById(comment.getId()).orElseThrow();
+        if (securityContext.isCallerInRole(UserRoles.ADMIN)
+                || originalComment.getUser().getUsername().equals(securityContext.getCallerPrincipal().getName())) {
+            comment.getUser().getComments().remove(originalComment);
+            commentRepository.delete(comment);
+        } else {
+            throw new EJBAccessException("Authorization failed for user " + securityContext.getCallerPrincipal().getName());
+        }
     }
 
     /**
@@ -76,8 +104,14 @@ public class CommentService implements Serializable {
      * @param id identifier of comment to update
      * @param comment data of updated comment
      */
-    public void updateComment(Long id, Comment comment) {
-        commentRepository.update(id, comment);
+    public void updateComment(Long id, Comment comment) throws EJBAccessException {
+        Comment originalComment = commentRepository.findById(id).orElseThrow();
+        if (securityContext.isCallerInRole(UserRoles.ADMIN)
+                || originalComment.getUser().getUsername().equals(securityContext.getCallerPrincipal().getName())) {
+            commentRepository.update(id, comment);
+        } else {
+            throw new EJBAccessException("Authorization failed for user " + securityContext.getCallerPrincipal().getName());
+        }
     }
 
     /**
